refactor: enhance ModalConfirmDelete and update localization for sector deletion
- Updated ModalConfirmDelete component to use keyup event for confirmation input, improving user experience.
- Added new localization entries for sector deletion confirmation in English, Spanish, and Portuguese.
- Refactored SectorCard to utilize the new localized message for delete confirmation, ensuring consistency across the application.
- Introduced SectorWorkingDaySection component to streamline workday management within the sector settings.
